Extracted and created separate file
for each Drawer component
