+++ src/components/drawer/drawer-body.jsx
@@ -1,0 +1,13 @@
+import { cn } from "@/utilities/functions";
+
+// Drawer body.
+const DrawerBody = ( { children, className, ...props } ) => {
+	return (
+		<div className={ cn( 'px-5 flex flex-col flex-1 overflow-y-auto', className ) } { ...props }>
+			{ children }
+		</div>
+	);
+};
+DrawerBody.displayName = 'Drawer.Body';
+
+export default DrawerBody;
